refactor: 공지사항 임시커밋
