+++ src/app/(route)/customer/(route)/notice/notice-info/[noticeInfoId]/_components/templates/NoticeMetaContainer.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+import NoticeInfoItemSkeleton from "../NoticeInfoItemSkeleton";
+import NoticeMeta from "../organisms/NoticeMeta";
+
+interface NoticeMetaContainerProps {
+  infoIsLoading: boolean;
+  infoIsError: boolean;
+  noticeInfoData: any;
+  numberId: number;
+}
+
+const NoticeMetaContainer = ({
+  infoIsLoading,
+  infoIsError,
+  noticeInfoData,
+  numberId,
+}: NoticeMetaContainerProps) => {
+  return (
+    <>
+      {infoIsLoading || infoIsError ? (
+        <NoticeInfoItemSkeleton />
+      ) : (
+        <NoticeMeta data={noticeInfoData} id={numberId} />
+      )}
+    </>
+  );
+};
+
+export default NoticeMetaContainer;
