+++ src/app/(route)/customer/_types/AdminChecker.ts
@@ -1,0 +1,5 @@
+export type AdminChecker = {
+  role: "USER" | "ADMIN" | undefined;
+}
+    
+// adminChecker
