+++ src/services/customer/Recommend/deleteNoticeRecommend.ts
@@ -1,6 +1,6 @@
 import { apiRequest } from "@/services/instant";
 
-const deleteNoticeRecommend = async (noticeId: number) => {
+const deleteNoticeRecommend = async (noticeId: string | string[]) => {
   const response = await apiRequest.delete(
     `api/recommend/notice/${noticeId}`
   );
